Made Custom Health Script
This health script can be placed on any object, enemy, or player that requires health. There's a neat custom inspector I made to display the health in a visual form live. I also made functions to deal damage and heal health, along with setting/getting the current phase of the unit.
This is a universal way to make health much easier to implement instead of hard coding health for each enemy. This is all visual.