Fixed Player and Plant Attack System
Now not only is it consistent, but we use the new health system I created to deal damage to the other opponent. This reduces dependencies and is visually cool too!
I also added a health bar field into the Health System so the player now uses that and I hide the current health phases if there are none.