Added Events to the health phases and the defeat
Now we can code custom logic for when a unit gets to a certain amount of HP and when they are defeated as well. If we don't put in any events, the game will resort to simply destroying the unit when HP is 0. With that, that wraps up the new health system!