Fixed Health Phase Bug and Added Icon
There was an error that occurred if we added a phase but didn't assign any events. It would be index out of range. I fixed it by first checking to see if  the persistent events  >0 and if so run the code.

I also added a heart icon for the Scriptable Objects so we can differentiate them easily in the assets folder